refactor(components/): Extract `handleFormSubmit`
